Introduce spec helpers
Why:
We would like to be able to have explicit tests that will allow to
ensure that prettier-ruby behaves as intended. Ideally these tests can
be run independently and could allow for TDD.

This commit:
introduces an assert_equal_and_valid method which takes an input string
and an expected output string and will validate that prettier can be
applied to the expected string by:

writing the expected string to a Tempfile, 'prettier-ruby-test.rb',
applying prettier-ruby to the Tempfile with the 'bin/print' executable,
asserting with minitest that the resultant file is equal to the expected
string,
and executing the resultant file with Open3 to capture any stderr and
status to validate that the formatted code can be run without throwing
an error.

Two example specs are provided in this commit to demonstrate the spec
helpers usage

Co-authored-by: christoomey <chris@thoughtbot.com>
Co-authored-by: johnschoeman <johnschoeman1617@gmail.com>