enable string symbol keys in hashes
For some reason, the parser classes `:'x'` as an `xstring`, but the key
part of { 'x': 1 } as a string. That also, for some reason, meant the
the `on_tstring_end` in the parser was not adding the relevant quote to
the sexp.

I added a very similar function to the parser, but for labels. All the
existing tests pass. I don't think there is a clean way of determining
when/where we can strip quotes from the keys, so I've had to disable the
linter to stop errors when running in hash rocket mode.

Fixes a regression from bf11b5e90753ca918baddee07a3c8020d8ef40d2
Closes #242