Always use lambda literal shorthand
The current output is invalid ruby syntax. I did
my best to read through the ruby grammar (parse.y)
and determine why a unary (!) would be valid in
->(...), but not in do |...| and I was grasping at
straws a bit. I'm not experienced enough with the
grammar there to determine if there's a bug in the
parser and it's picking the wrong matcher for args
in do |...| where a unary isn't allowed, but
nonetheless it breaks.

Example:

>> ->(active = !dead?) { '' } =>
lambda { |active = !dead?| '' } Traceback (most
recent call last): 3: from
/Users/jbielick/.rubies/ruby-2.6.6/bin/irb:23:in
`<main>' 2: from
/Users/jbielick/.rubies/ruby-2.6.6/bin/irb:23:in
`load' 1: from
/Users/jbielick/.rubies/ruby-2.6.6/lib/ruby/gems/2.6.0/gems/irb-1.0.0/exe/irb:11:in
`<top (required)>' SyntaxError ((irb):2: syntax
error, unexpected '!') lambda { |active = !dead?|
'' } ^ (irb):2: syntax error, unexpected '}',
expecting end-of-input lambda { |active = !dead?|
'' } Without determining exactly what the parsing
grammar allow and disallows, I found it difficult
to propose a change that would conditionally use
-> vs. lambda { when a particular arg pattern was
detected. It was my assumption that detection of
exceptions where -> must be used was futile. As a
result, I think the best course of action is to
use -> for everything, since the args patterns it
allows are a superset of do |...| (block params)
and it is likely unsafe to transform a ->(...)
into lambda { |...|.