maintain parser order during if/unless modifier
here's an example expression:

text = 'looooo...ng' if text

`text` is initialized in the current scope and is
now available as a variable. The formatting before
this commit produces:

if text
  text = 'loooo...ng'
end

The formatting is based on the fact that the expression
exceeds a line limit, so it breaks and uses the if..end
form. This has a nasty side effect in that text is not
initialized before it's referenced and produces a NameError.

https://docs.ruby-lang.org/en/2.0.0/syntax/control_expressions_rdoc.html#label-Modifier+if+and+unless

While the modifier and standard versions have both
a “test” expression and a “then” expression, they
are not exact transformations of each other due to
parse order. Here is an example that shows the
difference:

p a if a = 0.zero?  This raises the NameError
“undefined local variable or method `a'”.

When ruby parses this expression it first
encounters a as a method call in the “then”
expression, then later it sees the assignment to a
in the “test” expression and marks a as a local
variable.

When running this line it first executes the
“test” expression, a = 0.zero?.

Since the test is true it executes the “then”
expression, p a. Since the a in the body was
recorded as a method which does not exist the
NameError is raised.

The same is true for unless.