Trailing commas in hash arguments
Previously, printing this:

```ruby
patch '/api/v1/shops', params: { shop: { name: 'Single Lane', category: 'food_and_drinks' } }
```

would result in this:

```ruby
patch '/api/v1/shops',
      params: {
        shop: { name: 'Single Lane', category: 'food_and_drinks' }
      }
```

even if trailing commas was enabled. This is incorrect, as the third
line should have a trailing comma on it. The issue is that we were
attempting to handle trailing commas too low in the tree (from within
the assoclist_from_args node printing as opposed to the hash printing).