Add parens for necessary modifier statements
Occasionally you can end up in a situation where you can't use a
modifier form of an if/unless statement or a while/until loop without
parentheses. This includes if its inside a hash literal, array literal,
or its a part of an assignment. In these cases you need to wrap
everything in another set of parentheses.