Fix additional issues with AlpineJS syntax in haml
In PR https://github.com/prettier/plugin-ruby/pull/1097, I added support
for haml attributes starting with '@', which are extensively used
in AlpineJS.

However, the PR was incomplete, and did not work in all cases. This
PR adds two jest test cases that fail with the current code, and a fix
that passes all specs.

The fix also collocates @-attributes support with other haml edge cases
(in the printer) instead of hooking into a separate part of the
codebase.